+++ NEWS.md
@@ -1,3 +1,20 @@
+# rasterpic (development version)
+
+-   Add support for:
+
+    -   **terra**: `SpatVector`, `SpatExtent`,
+
+    -   **sf**: `sfg`, `sf::st_bbox()`.
+
+    -   Numeric vectors `c(xmin, ymin, xmax, ymax)`
+
+-   Now `img` parameter on `rasterpic_img()` accepts online images.
+
+-   Add an [article](https://dieghernan.github.io/rasterpic/articles/plots.html)
+    to the pkgdown site
+
+-   Fix CRAN errors.
+
 # rasterpic 0.1.0
 
 -   Initial release